requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.39.11
botocore==1.39.11
certifi==2025.8.3
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
influxdb-client==1.49.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
reactivex==4.0.4
s3transfer==0.13.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
